Working Beta
There may be a few bugs still:
When quitting and reopening OBS, the base color sources seems to revert to a previous size. I wonder if this is machine specific, or if it is somewhere hidden in the code.
When adding a HUD, the text sources will not update to the correct font, until the GUI is interacted with by the user. I have not found a way to force an update when adding the HUD.
This has not been throughly game tested

Features needed before its usable:
Dynamic text adjustment, which shrinks the font for long names

More color sources to break up the HUD into sections